+++ src/view/smart.js
@@ -1,0 +1,42 @@
+import Abstract from './abstract';
+
+export default class Smart extends Abstract {
+  constructor() {
+    super();
+    this._state = {};
+  }
+
+  updateData(update, justDataUpdating) {
+    if (!update) {
+      return;
+    }
+
+    this._state = Object.assign(
+      {},
+      this._state,
+      update,
+    );
+
+    if (justDataUpdating) {
+      return;
+    }
+
+    this.updateElement();
+  }
+
+  updateElement() {
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreHandlers();
+  }
+
+  restoreHandlers() {
+    throw new Error('Abstract method not implemented: resetHandlers');
+  }
+}
